requirements.txt
python-magic-bin==0.4.14
watchdog==6.0.0
